Remove RoundEntity and RoundStore because not used anymore
