+++ framework/src/modules/chain/components/storage/sql/accounts/reset_mem_tables.sql
@@ -20,7 +20,6 @@
 */
 
 DELETE FROM mem_accounts;
-DELETE FROM mem_round;
 DELETE FROM mem_accounts2delegates;
 DELETE FROM mem_accounts2multisignatures;
 DELETE FROM rounds_rewards;
