+++ framework/src/modules/chain/components/storage/sql/round_delegates/summed_round.sql
