--- framework/src/modules/chain/components/storage/sql/rounds/is_persisted.sql
@@ -1,17 +1,0 @@
-/*
- * Copyright © 2019 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- */
-
-SELECT EXISTS (
-	SELECT 1 FROM mem_round ${parsedFilters:raw}
-);
