--- framework/src/modules/chain/components/storage/sql/rounds/perform_votes_snapshot.sql
@@ -1,25 +1,0 @@
-/*
- * Copyright © 2019 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- */
-
-
-/*
-  DESCRIPTION: ?
-
-  PARAMETERS: ?
-*/
-
-CREATE TABLE mem_votes_snapshot AS
-SELECT address, "publicKey", vote, "producedBlocks", "missedBlocks"
-FROM mem_accounts
-WHERE "isDelegate" = 1
