--- framework/src/modules/chain/components/storage/sql/rounds/restore_round_snapshot.sql
@@ -1,22 +1,0 @@
-/*
- * Copyright © 2019 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- */
-
-
-/*
-  DESCRIPTION: ?
-
-  PARAMETERS: ?
-*/
-
-INSERT INTO mem_round SELECT * FROM mem_round_snapshot;
