--- framework/src/modules/chain/components/storage/sql/rounds/update_one.sql
@@ -1,15 +1,0 @@
-/*
- * Copyright © 2019 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- */
-
-UPDATE mem_round SET ${updateSet:raw} WHERE id = (SELECT id FROM mem_round ${parsedFilters:raw} LIMIT 1);
