+++ framework/src/modules/chain/transactions/transactions_handlers.js
@@ -19,7 +19,6 @@ const {
 	Status: TransactionStatus,
 	TransactionError,
 } = require('@liskhq/lisk-transactions');
-const votes = require('./votes');
 const votesWeight = require('./votes_weight');
 const exceptionsHandlers = require('./exceptions_handlers');
 const StateStore = require('../state_store');
@@ -131,7 +130,6 @@ const applyGenesisTransactions = storage => async (
 	const transactionsResponses = transactions.map(transaction => {
 		const transactionResponse = transaction.apply(stateStore);
 
-		votes.apply(stateStore, transaction);
 		votesWeight.apply(stateStore, transaction);
 		stateStore.transaction.add(transaction);
 
@@ -183,7 +181,6 @@ const applyTransactions = (storage, exceptions) => async (transactions, tx) => {
 			}
 
 			if (transactionResponse.status === TransactionStatus.OK) {
-				votes.apply(stateStore, transaction, exceptions);
 				votesWeight.apply(stateStore, transaction, exceptions);
 				stateStore.transaction.add(transaction);
 			}
@@ -284,7 +281,6 @@ const undoTransactions = (storage, exceptions) => async (
 
 	const transactionsResponses = transactions.map(transaction => {
 		const transactionResponse = transaction.undo(stateStore);
-		votes.undo(stateStore, transaction, this.exceptions);
 		votesWeight.undo(stateStore, transaction, this.exceptions);
 		return transactionResponse;
 	});
