+++ framework/test/jest/unit/specs/modules/chain/dpos/apply.spec.js
@@ -43,8 +43,6 @@ describe('dpos.apply()', () => {
 				},
 				RoundDelegates: {
 					getRoundDelegates: jest.fn().mockReturnValue(delegatePublicKeys),
-				},
-				Round: {
 					summedRound: jest.fn(),
 				},
 			},
@@ -207,7 +205,7 @@ describe('dpos.apply()', () => {
 				};
 			};
 
-			stubs.storage.entities.Round.summedRound.mockResolvedValue([
+			stubs.storage.entities.RoundDelegates.summedRound.mockResolvedValue([
 				{
 					fees: totalFee, // dividable to ACTIVE_DELEGATE count
 					rewards: delegatesWhoForged.map(() => rewardPerDelegate),
@@ -309,7 +307,7 @@ describe('dpos.apply()', () => {
 		it('should give the remainingFee ONLY to the last delegate of the round who forged', async () => {
 			// Arrange
 			const remainingFee = randomInt(5, 10);
-			stubs.storage.entities.Round.summedRound.mockResolvedValue([
+			stubs.storage.entities.RoundDelegates.summedRound.mockResolvedValue([
 				{
 					fees: totalFee + remainingFee,
 					rewards: delegatesWhoForged.map(() => rewardPerDelegate),
@@ -387,7 +385,7 @@ describe('dpos.apply()', () => {
 		describe('When all delegates successfully forges a block', () => {
 			it('should NOT update "missedBlocks" for anyone', async () => {
 				// Arrange
-				stubs.storage.entities.Round.summedRound.mockResolvedValue([
+				stubs.storage.entities.RoundDelegates.summedRound.mockResolvedValue([
 					{
 						fees: totalFee,
 						rewards: delegateAccounts.map(() => randomInt(1, 20)),
@@ -418,7 +416,9 @@ describe('dpos.apply()', () => {
 			it('should throw the error message coming from summedRound method and not perform any update', async () => {
 				// Arrange
 				const err = new Error('dummyError');
-				stubs.storage.entities.Round.summedRound.mockRejectedValue(err);
+				stubs.storage.entities.RoundDelegates.summedRound.mockRejectedValue(
+					err,
+				);
 
 				// Act && Assert
 				await expect(dpos.apply(lastBlockOfTheRound, stubs.tx)).rejects.toBe(
@@ -450,7 +450,7 @@ describe('dpos.apply()', () => {
 					randomInt(10, 1000),
 				);
 
-				stubs.storage.entities.Round.summedRound.mockResolvedValue([
+				stubs.storage.entities.RoundDelegates.summedRound.mockResolvedValue([
 					{
 						fees: totalFee.toString(), // dividable to ACTIVE_DELEGATE count
 						rewards: delegatesWhoForged.map(() => rewardPerDelegate),
