+++ framework/test/mocha/integration/process_transactions.js
@@ -26,7 +26,6 @@ const transactionsModule = require('../../../src/modules/chain/transactions');
 const {
 	TransactionInterfaceAdapter,
 } = require('../../../src/modules/chain/interface_adapters');
-const votes = require('../../../src/modules/chain/transactions/votes');
 
 const interfaceAdapters = {
 	transactions: new TransactionInterfaceAdapter(registeredTransactions),
@@ -346,16 +345,6 @@ describe('processTransactions', () => {
 							transactionStatus.PENDING,
 						);
 					});
-				});
-
-				it('should call votes.apply with correct params', async () => {
-					sinonSandbox.spy(votes, 'apply');
-					const { stateStore } = await applyTransactions(appliableTransactions);
-					expect(votes.apply.firstCall.args).to.be.eql([
-						stateStore,
-						appliableTransactions[0],
-						exceptions,
-					]);
 				});
 			});
 		});
